Refactor ShardingSphereSQLException
