+++ shardingsphere-error/shardingsphere-common-error/src/main/java/org/apache/shardingsphere/error/exception/lock/TableLockedException.java
@@ -15,19 +15,19 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.error.exception.standard.lock;
+package org.apache.shardingsphere.error.exception.lock;
 
-import org.apache.shardingsphere.error.exception.standard.ShardingSphereSQLException;
+import org.apache.shardingsphere.error.exception.ShardingSphereSQLException;
 import org.apache.shardingsphere.error.sqlstate.XOpenSQLState;
 
 /**
- * Table lock wait timeout exception.
+ * Table locked exception.
  */
-public final class TableLockWaitTimeoutException extends ShardingSphereSQLException {
+public final class TableLockedException extends ShardingSphereSQLException {
     
-    private static final long serialVersionUID = 2599713085782288003L;
+    private static final long serialVersionUID = 2622020743612706932L;
     
-    public TableLockWaitTimeoutException(final String schemaName, final String tableName, final long timeoutMilliseconds) {
-        super(XOpenSQLState.GENERAL_ERROR, 1301, "The table `%s` of schema `%s` lock wait timeout of %s ms exceeded", tableName, schemaName, String.valueOf(timeoutMilliseconds));
+    public TableLockedException(final String schemaName, final String tableName) {
+        super(XOpenSQLState.GENERAL_ERROR, 1302, "The table `%s` of schema `%s` is locked", tableName, schemaName);
     }
 }
