+++ shardingsphere-proxy/shardingsphere-proxy-frontend/shardingsphere-proxy-frontend-core/src/main/java/org/apache/shardingsphere/proxy/frontend/exception/CircuitBreakException.java
@@ -15,19 +15,19 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.error.exception.standard.lock;
+package org.apache.shardingsphere.proxy.frontend.exception;
 
-import org.apache.shardingsphere.error.exception.standard.ShardingSphereSQLException;
+import org.apache.shardingsphere.error.exception.ShardingSphereSQLException;
 import org.apache.shardingsphere.error.sqlstate.XOpenSQLState;
 
 /**
- * Table locked exception.
+ * Circuit break exception.
  */
-public final class TableLockedException extends ShardingSphereSQLException {
+public final class CircuitBreakException extends ShardingSphereSQLException {
     
-    private static final long serialVersionUID = 2622020743612706932L;
+    private static final long serialVersionUID = 6339672680026286798L;
     
-    public TableLockedException(final String schemaName, final String tableName) {
-        super(XOpenSQLState.GENERAL_ERROR, 1302, "The table `%s` of schema `%s` is locked", tableName, schemaName);
+    public CircuitBreakException() {
+        super(XOpenSQLState.GENERAL_WARNING, 1000, "Circuit break open, the request has been ignored");
     }
 }
